[ALPS09822132] iommu: change sec handle type
[Description]
Change sec handle type in iommu_test.

[Test]
Check whether phone can boot to HS.

MTK-Commit-Id: b7aa44bc7db3ce480bc4c07c19083f0bb9ae1ed2

CR-Id: ALPS09822132
Feature: iommu_kernel
(cherry picked from commit 8a044f822a8db36d4dc04ce4b36c2c4f14863893)

+++ drivers/misc/mediatek/iommu/iommu_test.c
@@ -686,7 +686,7 @@ static void dmabuf_map_iova(enum DMABUF_HEAP heap, int buf_id, int attach_id)
 	struct dma_buf_attachment *attach;
 	struct sg_table *table;
 #ifdef SECURE_HANDLE_TEST
-	u32 sec_handle;
+	u64 sec_handle;
 #endif
 	pr_info("%s start, heap:%s(%d), dev:%s\n", __func__,
 		heap_obj[heap].name, heap, dev_name(dev));
@@ -704,7 +704,7 @@ static void dmabuf_map_iova(enum DMABUF_HEAP heap, int buf_id, int attach_id)
 	if (!sec_handle)
 		pr_info("%s get secure handle failed, heap:%s\n", __func__, heap_obj[heap].name);
 	else
-		pr_info("%s get secure handle(%u) success, heap:%s\n", __func__,
+		pr_info("%s get secure handle(%llu) success, heap:%s\n", __func__,
 			sec_handle, heap_obj[heap].name);
 #endif
 	ret = check_attach_id(info_obj, attach_id);
